+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * [REMOVED] - `connect-flash` peer dependency. The module was being explicitly loaded during session initialization `Route.prototype.useSession`
 * [FIXED] - implementation of handling the `baseUrl` config value. Routers are now aware of the root base path value. This way we don't break API for modules which depend on it.
+* [ADDED] - http json response fallback when request fails and the custom `error-response` event listener is registered
 
 ## v0.11.1